pyproject.toml
[project]
name = "copo-do-mundo"
version = "0.1.0"
description = "Aplicação para competição de drinks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "seaborn>=0.13.2",
    "streamlit>=1.53.1",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-storage>=2.18.0",
    "google-auth>=2.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100

[tool.ruff.format]
exclude = ["*.ipynb"]

[tool.ruff.lint]
exclude = ["*.ipynb"]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

